Delete directory with now useless dummy file
This file was made so that PCH would work despite the existence of the file "ImageC.c" which could not be made into ".cpp" under the restrictions of libpng. Since then libpng, it has been replaced by libspng, which means the problematic file has been deleted, making this directory and its single file, completely useless.